--- app/Models/ResidentKey.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use App\Models\Abstract\KeyModel;
-
-class ResidentKey extends KeyModel
-{
-    protected static $recordModel = Resident::class;
-}
